+++ ThirdPartyAdapters/verizonmedia/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Verizon Media Android Mediation Adapter Changelog
 
+## Version 1.2.1.0
+- Support for Verizon Media SDK v1.2.1.
+- Support for Native ads added.
+- Support for Rewarded Video added.
+
 #### Version 1.1.1.0
 - Adapter has been rebranded to "Verizon Media".
 - Compatible with Verizon Media SDK 1.1.1.